Update colourOMatic.html
Waffy on Discord requested a new feature that I could provide, so I quickly added it! Now there is a checkbox that will count how many possible colours there are for two parents, not just show them. This can help with direct comparisons, and is how my future project is planned to work.